requirements.txt
python-can
crc16
pyserial
